Add dev SDL 1.2.15+ framework for Mac only
Note: the last commit is from 07.09.2016:
http://hg.libsdl.org/SDL/changeset/0c1a8b0429a9

+++ source/xcode/Frameworks/SDL.framework/Versions/A/Headers/SDL_endian.h
@@ -86,10 +86,10 @@ static __inline__ Uint16 SDL_Swap16(Uint16 x)
 #elif defined(__GNUC__) && (defined(__powerpc__) || defined(__ppc__))
 static __inline__ Uint16 SDL_Swap16(Uint16 x)
 {
-	Uint16 result;
+	int result;
 
 	__asm__("rlwimi %0,%2,8,16,23" : "=&r" (result) : "0" (x >> 8), "r" (x));
-	return result;
+	return (Uint16)result;
 }
 #elif defined(__GNUC__) && (defined(__m68k__) && !defined(__mcoldfire__))
 static __inline__ Uint16 SDL_Swap16(Uint16 x)
